+++ src/darkleaf/di/protocols.clj
@@ -10,14 +10,11 @@
 
 (ns darkleaf.di.protocols)
 
-(defprotocol Stoppable
-  (unwrap [this])
-  (stop [this]
-    "Stops the object. Returns nothing."))
-
 (defprotocol Factory
   (dependencies [this]
     "Returns a map of a key and a dependency type.
      A type can be `:required` or `:optional`.")
   (build [this dependencies]
-    "Builds a stoppable object from dependencies."))
+    "Builds an object from dependencies.")
+  (demolish [this obj]
+    "Demolishes or stops an object."))