+++ test/darkleaf/di/opt_ref_test.clj
@@ -3,17 +3,22 @@
    [clojure.test :as t]
    [darkleaf.di.core :as di]))
 
+(defn get-ref [{r   ::ref
+                :or {r :default}}]
+  r)
+
 (t/deftest opt-ref-test
-  (with-open [obj (di/start ::root
-                            {::root        (di/opt-ref ::replacement)
-                             ::replacement ::stub})]
-    (t/is (= ::stub @obj))))
+  (with-open [get-ref (di/start `get-ref
+                                {::ref         (di/opt-ref ::replacement)
+                                 ::replacement ::stub})]
+    (t/is (= ::stub (get-ref)))))
 
 (t/deftest opt-ref-missed-test
-  (with-open [obj (di/start ::root
-                            {::root (-> (di/opt-ref `dep)
-                                        (di/fmap (fnil identity :default)))})]
-    (t/is (= :default @obj))))
+  (with-open [get-ref (di/start `get-ref
+                                {::ref         (di/opt-ref ::replacement)
+                                 #_#_
+                                 ::replacement ::stub})]
+    (t/is (= :default (get-ref)))))
 
 (t/deftest pr-test
   (t/is (= "#darkleaf.di.core/opt-ref :darkleaf.di.opt-ref-test/object"