+++ test/darkleaf/di/tutorial/c_stop_test.clj
@@ -5,8 +5,7 @@
   {:nextjournal.clerk/visibility {:result :hide}}
   (:require
    [clojure.test :as t]
-   [darkleaf.di.core :as di]
-   [darkleaf.di.protocols :as dip]))
+   [darkleaf.di.core :as di]))
 
 ;; To stop a component, you should teach DI how to do it.
 ;; Use `::di/stop` to define a stop function.
@@ -31,19 +30,3 @@
 ;;     :or {max-conn 50}}]
 ;;   (ConnctionManager. max-conn))
 ;; ```
-
-;; You can also manually implement `dip/Stoppable` via `reify`, `extend-protocol`, etc.
-
-(defn root-explicit
-  [{::keys [*stopped?]}]
-  (reify dip/Stoppable
-    (unwrap [_]
-      *stopped?)
-    (stop [_]
-      (reset! *stopped? true))))
-
-(t/deftest stop-explicit-test
-  (let [*stopped? (atom false)]
-    (with-open [root (di/start `root-explicit {::*stopped? *stopped?})]
-      (t/is (= false @@root)))
-    (t/is @*stopped?)))